Scene now handles nodes
The scene now handles nodes.  This means that we can undo and redo actions such as node creation, deletion, movement and editing (unimplemented).

I have also allowed nodes to be selected, the grid to be toggled, the scene to be reset, and have cleaned up the node serialisation by moving it into cpp files.