Node editor
You can now edit the nodes using the UI.  You can also create new nodes by using the menubar or context menu in the scene tree.

I also made it so that when you have unsaved changes and would like to load a new scene or quit the application, the app prompts you to save before losing your changes.  This doesn't apply when you close the window or press Command-Q on Mac, for example.